Update RHINO_PLUG_IN_DESCRIPTION
Rhino doesn't correctly use the right code page for RHINO_PLUG_IN_DESCRIPTION causing malform characters when using two byte UTF-8 character encoding. So instead of seeing ©, it shows Â©.  This appears to be the only plugin that using two byte UTF-8 in the plugin description so the simplest solution is just to remove this character.